+++ dev/visualizer/Large Network Visualizer/src/sim_mob/vis/simultion/AgentTick.java
@@ -11,13 +11,13 @@ import sim_mob.vis.network.basic.ScaledPoint;
  * \author Zhang Shuai
  */
 public abstract class AgentTick implements DrawableAgent {
-	public AgentTick(int id) {
+	public AgentTick(long id) {
 		this.id = id;
 	}
 	
-	protected int id;
+	protected long id;
 	protected ScaledPoint pos;
 	public ScaledPoint getPos() { return pos; }
-	public int getID() { return id; }
+	public long getID() { return id; }
 	
 }
