+++ dev/visualizer/Large Network Visualizer/src/sim_mob/vis/simultion/DriverTick.java
@@ -88,7 +88,7 @@ public class DriverTick extends AgentTick {
 	 */
 
 
-	public DriverTick(int id, double posX, double posY, double angle) {
+	public DriverTick(long id, double posX, double posY, double angle) {
 		this(id, posX, posY, angle, null);
 	}
 	
@@ -111,7 +111,7 @@ public class DriverTick extends AgentTick {
 	}
 	
 	
-	public DriverTick(int id, double posX, double posY, double angle, RxLocation msgLocation) {
+	public DriverTick(long id, double posX, double posY, double angle, RxLocation msgLocation) {
 		super(id);
 		
 		this.pos = new FlippedScaledPoint(posX, posY);
@@ -361,7 +361,7 @@ public class DriverTick extends AgentTick {
 		g.setFont(idFont);
 		g.setStroke(new BasicStroke(0.5F));
 		
-		String id = Integer.toString(getID());
+		String id = Long.toString(getID());
 		g.drawString(id, 0, 0);
 
 		//Restore AffineTransform matrix.
