+++ dev/visualizer/Large Network Visualizer/src/sim_mob/vis/simultion/SignalLineTick.java
@@ -16,15 +16,15 @@ public class SignalLineTick{
 	private long id;
 	private ArrayList<ArrayList<Integer>> allVehicleLights;
 	private ArrayList<Integer> allPedestrianLights;	
-	private Integer intersectionID;
+	private Long intersectionID;
 //	private Phase [] phases; 
 	private boolean fake;
 	
-	public Integer getIntersectionID(){ return intersectionID;}
+	public Long getIntersectionID(){ return intersectionID;}
 	public ArrayList<ArrayList<Integer>> getVehicleLights(){ return allVehicleLights;}
 	public ArrayList<Integer> getPedestrianLights(){return allPedestrianLights;}
 	public boolean getFake(){ return fake; }
-	public SignalLineTick(long id, ArrayList<ArrayList<Integer>> allVehicleLights, ArrayList<Integer> allPedestrainLights,Integer intersectionID){
+	public SignalLineTick(long id, ArrayList<ArrayList<Integer>> allVehicleLights, ArrayList<Integer> allPedestrainLights,Long intersectionID){
 		this.id = id;
 		this.allVehicleLights = allVehicleLights;
 		this.allPedestrianLights =  allPedestrainLights;
@@ -43,12 +43,12 @@ public class SignalLineTick{
 	//my solution:
 	private Hashtable<String ,ArrayList<TrafficSignalLine>> TrafficSignalLines;	//String is phase (A,B,C,...)
 	private HashMap<TrafficSignalLine, Color> TrafficSignalLines_Map;
-	private HashMap<Integer, Integer> CrossingID_Map;
+	private HashMap<Long, Integer> CrossingID_Map;
 	public Hashtable<String ,ArrayList<TrafficSignalLine>> getAllTrafficSignalLines(){ return TrafficSignalLines;}
 	public HashMap<TrafficSignalLine, Color> getAllTrafficSignalLines_Map(){ return TrafficSignalLines_Map;}
-	public HashMap<Integer, Integer> getCrossingID_Map(){ return CrossingID_Map;}
+	public HashMap<Long, Integer> getCrossingID_Map(){ return CrossingID_Map;}
 
-	public SignalLineTick(int id,ArrayList<TrafficSignalLine> TrafficSignalLines_,HashMap<Integer, Integer> CrossingIDs_map,int tempTick_, String tempPhase_){
+	public SignalLineTick(long id,ArrayList<TrafficSignalLine> TrafficSignalLines_,HashMap<Long, Integer> CrossingIDs_map,int tempTick_, String tempPhase_){
 		this.id = id;
 		//at present we leverage the similarity of signal id and intersectionId.
 		//In case one day they decided them to be different,this method should 
@@ -69,8 +69,8 @@ public class SignalLineTick{
 			TrafficSignalLines_Map.put(tsl, tsl.getCurrColor());
 		}
 		
-		CrossingID_Map = new HashMap<Integer, Integer>();
-		for(Integer i:CrossingIDs_map.keySet())
+		CrossingID_Map = new HashMap<Long, Integer>();
+		for(Long i:CrossingIDs_map.keySet())
 		{
 //			if((tempTick == 230)&& tempPhase.equals("D"))
 //				System.out.println("Tick 230 Phase D Setting crossing " + i + "  color to " + CrossingIDs_map.get(i));
