+++ dev/visualizer/Large Network Visualizer/src/sim_mob/vis/simultion/TimeTick.java
@@ -10,22 +10,22 @@ import java.util.Hashtable;
  * \author Zhang Shuai
  */
 public class TimeTick {
-	public Hashtable<Integer, AgentTick> agentTicks;
+	public Hashtable<Long, AgentTick> agentTicks;
 	
 	/**
 	 * Tracking ticks are agents of any type that "track" an agent with the same
 	 * ID in agentTicks. 
 	 */
-	public Hashtable<Integer, AgentTick> trackingTicks;
+	public Hashtable<Long, AgentTick> trackingTicks;
 	
 	//public Hashtable<Integer, SignalTick> signalTicks;
-	public Hashtable<Integer, SignalLineTick> signalLineTicks;
+	public Hashtable<Long, SignalLineTick> signalLineTicks;
 	
 	
 	public TimeTick() {
-    	trackingTicks = new Hashtable<Integer, AgentTick>();
-    	agentTicks = new Hashtable<Integer, AgentTick>();
+    	trackingTicks = new Hashtable<Long, AgentTick>();
+    	agentTicks = new Hashtable<Long, AgentTick>();
     	//signalTicks = new Hashtable<Integer, SignalTick>();
-    	signalLineTicks = new Hashtable<Integer,SignalLineTick>();
+    	signalLineTicks = new Hashtable<Long,SignalLineTick>();
 	}
 }
